Rework calculation of curve's x and y coordinates
Precomputing all the x and y coordinates and storing them into an array
is problematic for several reasons. First, the array takes up a lot of
space. Furthermore, if there are more than 2^31 - 1 pairs of
coordinates to store in the array, then the array cannot even be
instantiated. Thirdly, mapping the t parameter to an index in the
array was complicated due to imprecision in floating-point arithmetic.
We could use a HashMap to remedy this particular mapping issue, but the
HashMap suffers from the same space problem as the array. Finally,
there is no noticeable performance difference between precomputation
and computing the x and y coordinates as the program runs. For these
reasons, we choose to compute the x and y coordinates as the curve is
being traced out.

In addition, the t parameter is now guaranteed to stay within the
ButterflyCurve's tLowerBound and tUpperBound (so now the lower and
upper bounds are truly bounds).