Create EquationParser and implement some methods
The EquationParser's purpose is to evaluate mathematical expressions
containing numbers, operators, trigonometric functions, and a variable
"t". This commit fully implements the bodies of the methods evaluate()
and tokenize(). Also, unit tests have been written for tokenize(), but
there are no tests for evaluate() yet.

In general, an expression is evaluated in three steps. First, the
expression is parsed into a list of tokens. Then, the Shunting-Yard
algorithm is run on the list of tokens to produce a queue of tokens
representing the expression written in postfix notation. Finally, the
postfix expression is evaluated. EquationParser has one method for
every step.