Add custom parametric equation input; bug fixes
The program now graphs custom parametric equations input by the user.
The program allows equations containing the operators +, -, *, /, ^, (,
and ), as well as the trigonometric functions sin, cos, and tan.

The UI has changed; the window is now divided into an EquationPanel
(where the user inputs their parametric equations and the bounds for
the t variable) and a DrawingPanel (where the parametric curve input
by the user is plotted).

Some bugs in EquationParser have been fixed, and more unit tests for
EquationParser have been added to achieve better code coverage.